requirements.txt
boto3==1.25.5
botocore==1.28.5
certifi==2022.9.24
charset-normalizer==2.1.1
emoji==2.1.0
greenlet==1.1.3.post0
idna==3.4
jmespath==1.0.1
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.42
urllib3==1.26.12
